+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 ### Features
+
 - `[jest-cli]`: Add `compact` option to default reporter ([#7845](https://github.com/facebook/jest/pull/7845))
 
 ### Fixes