+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Chore & Maintenance
 
-- `[*]` Make sure to copy over `d.ts` files to the `build` directory when building ([#8086](https://github.com/facebook/jest/pull/8086))
+- `[*]` Make sure to include `d.ts` files in the tarball when building ([#8086](https://github.com/facebook/jest/pull/8086))
 
 ### Performance